printf funcional

+++ ft_print.h
@@ -6,7 +6,7 @@
 /*   By: ide-dieg <ide-dieg@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/28 17:04:17 by ide-dieg          #+#    #+#             */
-/*   Updated: 2024/02/28 18:53:24 by ide-dieg         ###   ########.fr       */
+/*   Updated: 2024/02/29 20:08:21 by ide-dieg         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 int	ft_putstr(char *s);
 int	ft_strlen(const char *str);
 int	ft_printf(char const *f, ...);
-int	ft_putnbr_base(int n, char *base, int lenbase);
+int	ft_putnbr_base(long n, char *base, int lenbase);
 int	ft_putchar(char c);
 
 #endif
